+++ CHANGES.rst
@@ -17,6 +17,8 @@ Revision 0.2.9, released XX-10-2019
 - Fixed the issue of not operational `allowmultioptions` option
   when the statement spans multiple lines
 - Fixed the issue of key-only statements parsing
+- Added `disableemptyelementstags` option to disable parsing of
+  empty element tags.
 
 Revision 0.2.8, released 29-12-2018
 -----------------------------------